src:CodelTime class, Reflow of CoDelTimestampTag and QueueTimestampTag classes and various renamings

+++ src/internet/model/fq_codel-queue.h
@@ -78,7 +78,7 @@ private:
   UniformVariable m_psource;
   mutable uint32_t m_peturbation;
   uint32_t m_quantum;
-  uint32_t m_backlog;
+  uint64_t m_backlog;
 };
 
 } // namespace ns3
